Add files via upload
Added all 3 here, not sure how to add to readme. 
what was hard about this project was figuring out the formatting. I was testing it on an Iphone 15, but the app only allows to test on an Iphone X/11, so there were some sizing issues and formatting problems, which I had to mess around with. I eventually came up on the "Fill parent" for the vertical arrangements.etc, which helped me format everything.